Add test for Address Service

+++ src/main/java/com/hb/acadia/service/RoleService.java
@@ -5,11 +5,15 @@ import java.util.List;
 import javax.transaction.Transactional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.web.access.ExceptionTranslationFilter;
 import org.springframework.stereotype.Service;
+import org.springframework.web.servlet.mvc.method.annotation.ExceptionHandlerExceptionResolver;
 
 import com.hb.acadia.model.user.Role;
 import com.hb.acadia.model.user.User;
 import com.hb.acadia.repository.RoleRepository;
+
+import lombok.extern.slf4j.Slf4j;
 
 /**
  * Service class
@@ -17,23 +21,25 @@ import com.hb.acadia.repository.RoleRepository;
  * @author simonaliotti
  *
  */
+@Slf4j
 @Service
 public class RoleService {
 
 	@Autowired
-	RoleRepository roleRepository;
+	private RoleRepository roleRepository;
 	@Autowired
-	UserService userService;
+	private UserService userService;
 
-	
 	/**
 	 * Method to get a role by roleName
+	 * 
 	 * @param roleName
 	 * @return the given role
 	 */
 	public Role getRoleByRoleName(String roleName) {
 		return roleRepository.findRoleByRoleName(roleName);
 	}
+
 	/**
 	 * Method to get all roles
 	 * 
@@ -52,30 +58,35 @@ public class RoleService {
 	@Transactional
 	public Role createRole(Role role) {
 		return roleRepository.save(role);
-		
+
 	}
 
 	/**
-	 * Delete a role
-	 * Delete all roles. Set users's role to null before deleting role 
+	 * Delete a role Delete all roles. Set users's role to null before deleting role
+	 * 
 	 * @param role
+	 * @throws IllegalAccessException
 	 */
 	@Transactional
-	public void deleteRole(Role role) {
+	public void deleteRole(Role role) throws IllegalAccessException {
 		List<User> users = userService.getUserByRole(role);
-		users.forEach(user -> {
-			user.setRole(null);
-			userService.updateUser(user);
-		});
-		roleRepository.delete(role);
+		if (users.isEmpty()) {
+			roleRepository.delete(role);
+		} else {
+			throw new IllegalAccessException();
+		}
+
 	}
 
 	/**
-	 * Delete all roles. 
-	 * the roles
+	 * Delete all roles. the roles
+	 * 
+	 * @throws IllegalAccessException
 	 */
 	@Transactional
-	public void deleteAll() {
-		getRoles().forEach(role -> deleteRole(role));
+	public void deleteAll() throws IllegalAccessException {
+		for (Role role : getRoles()) {
+			deleteRole(role);
+		}
 	}
 }
